pyproject.toml
[project]
name = "az-ai"
version = "0.1.0"
description = "AZ AI CLI"
authors = [{name = "Dominique Broeglin", email = "dominique.broeglin@microsoft.com"}]
license = {file="LICENSE"}
readme = "README.md"
requires-python = ">= 3.11"
dependencies = [
    "azure-cli~=2.70.0",
    "python-dotenv ~=1.0.1",
    "typer>=0.15.2",
]

[project.scripts]
"az-ai" = "az_ai.cli:app"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.uv]
prerelease = "allow"
dev-dependencies = [
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
]

[tool.pytest.ini_options]
addopts = "--cov-branch --cov-report term --cov=core"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
markers=["unit: Unit tests.",
         "system: System (browser-based) tests."]
         
[tool.coverage.run]
source = ["core"]

[tool.ruff]
line-length = 120
fix = true
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
extend-fixable = ["E", "F", "UP", "B", "SIM", "I"]
